Refactor RGB plotting into separate files
