+++ Code/cyclegan_website/utils/createPlot.ts
@@ -1,0 +1,35 @@
+import * as d3 from "d3";
+
+export const createPlot = (
+  id: string,
+  width: number,
+  height: number,
+  margin: { top: any; right: any; bottom: any; left: any },
+  xlabel: string,
+  ylabel: string
+) => {
+  const svg = d3
+    .select(id)
+    .append("svg")
+    .attr("width", width + margin.left + margin.right)
+    .attr("height", height + margin.top + margin.bottom)
+    .append("g")
+    .attr("transform", "translate(" + margin.left + "," + margin.top + ")");
+
+  svg
+    .append("text")
+    .attr("text-anchor", "end")
+    .attr("x", width)
+    .attr("y", height + margin.top + 20)
+    .text(xlabel);
+
+  svg
+    .append("text")
+    .attr("text-anchor", "end")
+    .attr("transform", "rotate(-90)")
+    .attr("y", -margin.left + 20)
+    .attr("x", -margin.top)
+    .text(ylabel);
+
+  return svg;
+};
